ポケモン詳細画面を追加
